rename asset info to asset layer
