+++ python_modules/dagster/dagster/core/definitions/job_definition.py
@@ -38,7 +38,7 @@ from dagster.core.utils import str_format_set
 
 from .asset_layer import AssetLayer
 from .config import ConfigMapping
-from .assets_info import AssetsJobInfo
+from .asset_layer import AssetLayer
 from .executor_definition import ExecutorDefinition
 from .graph_definition import GraphDefinition, SubselectedGraphDefinition
 from .hook_definition import HookDefinition
@@ -73,7 +73,6 @@ class JobDefinition(PipelineDefinition):
         hook_defs: Optional[AbstractSet[HookDefinition]] = None,
         op_retry_policy: Optional[RetryPolicy] = None,
         version_strategy: Optional[VersionStrategy] = None,
-        assets_info: Optional[AssetsJobInfo] = None,
         _op_selection_data: Optional[OpSelectionData] = None,
         asset_layer: Optional[AssetLayer] = None,
     ):
@@ -90,9 +89,6 @@ class JobDefinition(PipelineDefinition):
         self._cached_partition_set: Optional["PartitionSetDefinition"] = None
         self._op_selection_data = check.opt_inst_param(
             _op_selection_data, "_op_selection_data", OpSelectionData
-        )
-        self._assets_info = check.opt_inst_param(
-            assets_info, "assets_nfo", AssetsJobInfo, default=AssetsJobInfo.from_graph(graph_def)
         )
 
         super(JobDefinition, self).__init__(
@@ -138,10 +134,6 @@ class JobDefinition(PipelineDefinition):
     @property
     def loggers(self) -> Mapping[str, LoggerDefinition]:
         return self.get_mode_definition().loggers
-
-    @property
-    def assets_info(self) -> AssetsJobInfo:
-        return self._assets_info
 
     def execute_in_process(
         self,
@@ -202,7 +194,7 @@ class JobDefinition(PipelineDefinition):
             tags=self.tags,
             op_retry_policy=self._solid_retry_policy,
             version_strategy=self.version_strategy,
-            assets_info=self.assets_info,
+            asset_layer=self.asset_layer,
         ).get_job_def_for_op_selection(op_selection)
 
         tags = None
@@ -267,8 +259,6 @@ class JobDefinition(PipelineDefinition):
             op_retry_policy=self._solid_retry_policy,
             graph_def=sub_graph,
             version_strategy=self.version_strategy,
-            # TODO: subset this structure
-            assets_info=self.assets_info,
             _op_selection_data=OpSelectionData(
                 op_selection=op_selection,
                 resolved_op_selection=set(
@@ -331,7 +321,7 @@ class JobDefinition(PipelineDefinition):
             hook_defs=hook_defs | self.hook_defs,
             description=self._description,
             op_retry_policy=self._solid_retry_policy,
-            assets_info=self.assets_info,
+            asset_layer=self.asset_layer,
             _op_selection_data=self._op_selection_data,
         )
