+++ python_modules/dagster/dagster/core/execution/plan/execute_step.py
@@ -430,26 +430,11 @@ def _asset_key_and_partitions_for_output(
     manager_asset_key = output_manager.get_output_asset_key(output_context)
 
     pipeline_def = output_context.step_context.pipeline_def
-    if isinstance(pipeline_def, JobDefinition):
-        node_handle = output_context.step_context.solid_handle
-        output_asset_info = pipeline_def.assets_info.asset_info_for_output(
-            node_handle=output_context.step_context.solid_handle, output_name=output_def.name
-        )
-        if output_asset_info:
-            if manager_asset_key is not None:
-                raise DagsterInvariantViolationError(
-                    f'The IOManager of output "{output_def.name}" on node "{node_handle}" associates it '
-                    f'with asset key "{manager_asset_key}", but this output has already been defined to '
-                    f'produce asset "{output_asset_info.asset_key}", either via a Software Defined Asset, '
-                    "or by setting the asset_key parameter on the OutputDefinition. In most cases, this "
-                    "means that you should use an IOManager that does not specify an AssetKey in its "
-                    "get_output_asset_key() function for this output."
-                )
-            return (
-                output_asset_info.asset_key,
-                output_asset_info.asset_partitions_fn(output_context) or set(),
-            )
-    elif output_def.is_asset:
+    node_handle = output_context.step_context.solid_handle
+    output_asset_info = pipeline_def.asset_layer.asset_info_for_output(
+        node_handle=output_context.step_context.solid_handle, output_name=output_def.name
+    )
+    if output_asset_info:
         if manager_asset_key is not None:
             raise DagsterInvariantViolationError(
                 f'The IOManager of output "{output_def.name}" on node "{node_handle}" associates it '
